updated the simple interest formula

+++ python_introduction/simple_interest.py
@@ -5,5 +5,5 @@ time = 3
 P = principal
 R = rate
 T = time
-interest = float(P * R * T)
+interest = P * R * T
 print("The simple interest is:", interest)
